# docs/openapi.yaml
openapi: 3.0.3
info:
  title: sooqa API
  version: 0.1.0
  description: Private single-admin HTTP API for the durable sooqa media pipeline.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
paths:
...
  /api/v1/ingest-requests:
    post:
      tags: [ingest]
      operationId: createIngestRequest
      summary: Submit a URL for durable ingestion
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestCreateRequest'
      responses:
        '202':
          description: Ingest accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestAcceptedResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/ingest-requests/{id}:
    get:
      tags: [ingest]
      operationId: getIngestRequest
      summary: Read ingest progress
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IngestId'
      responses:
        '200':
          description: Ingest request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: sooqa-api-token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 255 }
    IngestId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Idempotency or state conflict.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    PayloadTooLarge:
      description: Request body exceeds the configured limit.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalError:
      description: Unexpected server failure.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id, details]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            details: { type: object, additionalProperties: true }
...
    IngestCreateRequest:
      type: object
      required: [url]
      additionalProperties: false
      properties:
        url: { type: string, format: uri }
        page_url: { type: string, format: uri, nullable: true }
        page_title: { type: string, nullable: true }
        selected_text: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
    IngestAcceptedResponse:
      type: object
      required: [id, status, links]
      properties:
        id: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/IngestStatus' }
        links:
          type: object
          required: [self]
          properties:
            self: { type: string }
    IngestResponse:
      allOf:
        - $ref: '#/components/schemas/IngestAcceptedResponse'
        - type: object
          required: [kind, source_url, created_at, updated_at, supplied_tags]
          properties:
            kind: { type: string, enum: [url, telegram_message, upload] }
            source_url: { type: string }
            page_url: { type: string, nullable: true }
            page_title: { type: string, nullable: true }
            supplied_caption: { type: string, nullable: true }
            supplied_tags: { type: array, items: { type: string } }
            error_code: { type: string, nullable: true }
            error_message: { type: string, nullable: true }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
            completed_at: { type: string, format: date-time, nullable: true }
    IngestStatus:
      type: string
      enum: [received, queued, downloading, probing, exact_dedup_check, normalizing, fingerprinting, similarity_check, storing, completed, failed_retryable, failed_terminal, cancelled]
...